isolate deck layers hook
